MOD: moving method of moving object

+++ src/main/java/hu/oe/nik/szfmv/environment/object/IDriveable.java
@@ -11,8 +11,8 @@ public interface IDriveable {
 
     void turn(float angle);
 
-    void accelerate(int scale);
+    void accelerate(double scale);
 
-    void slow(int scale);
+    void slow(double scale);
 
 }
